# frontend/openapi/schemas/be-schema.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/be-nest-stocks/api/filters": {
            "post": {
                "operationId": "FilterController_create",
                "summary": "Create a new filter",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "The filter data to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFilterDto"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "A sample filter creation payload",
                                    "value": {
                                        "name": "Sample Filter",
                                        "numberCriteria": [
                                            {
                                                "property": "price",
                                                "condition": "less_than",
                                                "value": 100
                                            }
                                        ],
                                        "stringCriteria": [
                                            {
                                                "property": "exchange",
                                                "condition": "equal",
                                                "value": "NYSE"
                                            }
                                        ],
                                        "ratioCriteria": [
                                            {
                                                "numerator": "marketCap",
                                                "denominator": "revenuePerShare",
                                                "condition": "less_than",
                                                "value": 2
                                            }
                                        ],
                                        "user": "60b8d295f1dfb862d8f51d3b"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The filter has been successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Filter"
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            },
            "get": {
                "operationId": "FilterController_findAll",
                "summary": "Get all filters",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Return all filters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Filter"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            }
        },
        "/be-nest-stocks/api/filters/{id}": {
            "get": {
                "operationId": "FilterController_findOne",
                "summary": "Get a filter by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the filter to retrieve",
                        "example": "60b8d295f1dfb862d8f51d3b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the filter with the given ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Filter"
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            },
            "put": {
                "operationId": "FilterController_update",
                "summary": "Update a filter by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the filter to update",
                        "example": "60b8d295f1dfb862d8f51d3b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "The updated filter data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFilterDto"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "A sample filter update payload",
                                    "value": {
                                        "name": "Updated Filter",
                                        "numberCriteria": [
                                            {
                                                "property": "price",
                                                "condition": "less_than",
                                                "value": 200
                                            }
                                        ],
                                        "stringCriteria": [
                                            {
                                                "property": "name",
                                                "condition": "not_equal",
                                                "value": "Some Name"
                                            }
                                        ],
                                        "ratioCriteria": [
                                            {
                                                "numerator": "marketCap",
                                                "denominator": "revenuePerShare",
                                                "condition": "less_than",
                                                "value": 1
                                            }
                                        ],
                                        "user": "60b8d295f1dfb862d8f51d3b"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The filter has been successfully updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Filter"
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            },
            "delete": {
                "operationId": "FilterController_remove",
                "summary": "Delete a filter by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the filter to delete",
                        "example": "60b8d295f1dfb862d8f51d3b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The filter has been successfully deleted."
                    }
                },
                "tags": ["stocks-filters"]
            }
        },
        "/be-nest-stocks/api/filters/user/{userId}": {
            "get": {
                "operationId": "FilterController_findByUser",
                "summary": "Get filters by user ID",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "description": "User ID to retrieve filters for",
                        "example": "60b8d295f1dfb862d8f51d3b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the filters for the given user ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Filter"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            }
        },
        "/be-nest-stocks/api/filters/{id}/apply": {
            "get": {
                "operationId": "FilterController_applyFilter",
                "summary": "Apply a filter by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the filter to apply",
                        "example": "60b8d295f1dfb862d8f51d3b",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the stocks matching the filter criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Stock"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            }
        },
        "/be-nest-stocks/api/filters/apply-direct": {
            "post": {
                "operationId": "FilterController_applyFilterDirectly",
                "summary": "Apply a filter directly from the payload",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "The filter data to apply directly",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFilterDto"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "A sample filter apply directly payload",
                                    "value": {
                                        "name": "Temporary Filter",
                                        "numberCriteria": [
                                            {
                                                "property": "price",
                                                "condition": "greater_than",
                                                "value": 150
                                            }
                                        ],
                                        "stringCriteria": [
                                            {
                                                "property": "exchange",
                                                "condition": "equal",
                                                "value": "NASDAQ"
                                            }
                                        ],
                                        "ratioCriteria": [
                                            {
                                                "numerator": "marketCap",
                                                "denominator": "revenuePerShare",
                                                "condition": "greater_than",
                                                "value": 1.5
                                            }
                                        ],
                                        "user": "60b8d295f1dfb862d8f51d3b"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return the stocks matching the filter criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Stock"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["stocks-filters"]
            }
        },
...
    },
    "info": {
        "title": "Stock Screener API",
        "description": "API for stock screener application",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateFilterDto": {
                "type": "object",
                "properties": {}
            },
            "Filter": {
                "type": "object",
                "properties": {}
            },
            "Stock": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}